move thumbnails Python file to core
