+++ backend/app/core/thumbnails.py
